issue 8: edition of items + some bug fixes. (for example, you could not uncheck shopping list items)

+++ composeApp/src/commonMain/kotlin/com/example/foodmanager/data/MockDb.kt
@@ -265,7 +265,7 @@ object MockDb {
     }
 
 
-    fun deleteShoppingItem(item_id: Int) {
+    fun deleteShoppingItem(item_id: Long) {
         _shoppingitems.value = _shoppingitems.value.filter { it.id != item_id } // Delete elements with the same id
     }
